Refactoring to a more complete CDK implementation
